Clean up code and import more content
